backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
face-recognition==1.4.0
numpy==1.24.3
Pillow==10.0.0
opencv-python==4.8.0.76
python-dotenv==1.0.0
